Match HTML data with our data_set
We want to create a transformation script which reads URL
from a CSV file, then fetches the content of the webpage at
the URL.

We want to match and confirm that data scraped from the actual
webpage is consistent with out CSV dataset.

This is a trivial implementation of a transformation script. It
needs to be improved overtime.